requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
pillow>=10.2.0
numpy>=1.26.0
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
albumentations>=1.4.0
opencv-python>=4.9.0
scikit-learn>=1.4.0
scipy>=1.11.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pydantic>=2.5.0